+++ Assets/Scripts/Product Data.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Product Data")]
 public class ProductData : ScriptableObject
 {
-    public float productQuantity = 0;
+    public double productQuantity = 0;
 
     //Upgradables of product itself
-    public float productSpentPerService = 1f;
+    public double productSpentPerService = 1f;
     public int productCapacity;
 
 }
